Refactor batch usage stats data types and rounding
- Updated TimeSeriesDataPoint and StatsSummary models to use int for duration and response time
- Added rounding for avg_duration_ms, average_response_time, and cache_hit_rate in Elasticsearch queries
- Improved precision and type consistency for statistical metrics

+++ src/api/endpoints/batches.py
@@ -91,7 +91,7 @@ class TimeSeriesDataPoint(BaseModel):
     total_tokens: int
     prompt_tokens: int
     completion_tokens: int
-    avg_duration_ms: float
+    avg_duration_ms: int
     tokens_per_second: float
 
 
@@ -102,7 +102,7 @@ class StatsSummary(BaseModel):
     cached_tasks: int
     total_tokens: int
     completion_tokens: int
-    average_response_time: float
+    average_response_time: int
     tokens_per_second: float
     cache_hit_rate: float
